Outsource state Cards to App.js
